+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -626,231 +626,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal(onOptionChangedCallCount, 1, 'onOptionChanged call count');
     });
 
-    QUnit.test('Columns hiding - columnHidingEnabled is true', function(assert) {
-        // arrange
-        $('#container').width(200);
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columnHidingEnabled: true,
-            dataSource: [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy' }],
-            columns: ['firstName', 'lastName']
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-        const adaptiveColumnsController = instance.getController('adaptiveColumns');
-        let $visibleColumns;
-
-        this.clock.tick();
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-
-        // act
-        assert.equal($visibleColumns.length, 3, 'only 1 column is visible');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
-        assert.ok(dataGridWrapper.headers.isColumnHidden(1), 'second column is hidden');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is shown');
-        assert.equal($visibleColumns.eq(0).text(), 'First Name', 'it is \'firstName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'lastName', '\'lastName\' column is hidden');
-
-        $('#container').width(450);
-        instance.updateDimensions();
-        this.clock.tick();
-
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-
-        // assert
-        assert.equal($visibleColumns.length, 3, '2 columns are visible');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
-        assert.ok(dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is hidden');
-        assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
-        assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There is no hidden columns');
-        assert.equal(adaptiveColumnsController.getHidingColumnsQueue().length, 2, 'There is 2 columns in hiding queue');
-    });
-
-    QUnit.test('Columns hiding - hidingPriority', function(assert) {
-        // arrange
-        $('#container').width(200);
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy' }],
-            columns: [{ dataField: 'firstName', hidingPriority: 0 }, { dataField: 'lastName', hidingPriority: 1 }]
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-        const adaptiveColumnsController = instance.getController('adaptiveColumns');
-        let $visibleColumns;
-
-        this.clock.tick();
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-        const $hiddenColumn = $('.dx-datagrid-hidden-column').eq(0);
-
-        // act
-        assert.ok(dataGridWrapper.headers.isColumnHidden(0), 'first column is hidden');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is shown');
-        assert.equal($visibleColumns.length, 3, 'only 1 column is visible');
-        assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'it is \'lastName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'firstName', '\'firstName\' column is hidden');
-        // T824145
-        if(browser.msie || browser.chrome) {
-            assert.equal(parseInt($hiddenColumn.css('border-right-width')), 0, 'no right border');
-            assert.equal(parseInt($hiddenColumn.css('border-left-width')), 0, 'no left border');
-        }
-
-        $('#container').width(450);
-        instance.updateDimensions();
-        this.clock.tick();
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-
-        // assert
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
-        assert.ok(dataGridWrapper.headers.isColumnHidden(2), 'adaptive column is hidden');
-        assert.equal($visibleColumns.length, 3, '2 columns are visible');
-        assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
-        assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There is no hidden columns');
-        assert.equal(adaptiveColumnsController.getHidingColumnsQueue().length, 2, 'There is 2 columns in hiding queue');
-    });
-
-    QUnit.test('Columns hiding - column without priority must stay (hidingPriority)', function(assert) {
-        // arrange
-        $('#container').width(80);
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy van Dyk', age: 40, country: 'India' }],
-            columns: [{ dataField: 'firstName', hidingPriority: 0 }, { dataField: 'lastName', hidingPriority: 1 }, 'age', 'country']
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-        const adaptiveColumnsController = instance.getController('adaptiveColumns');
-        let $visibleColumns;
-
-        this.clock.tick();
-
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-
-        // act
-        assert.ok(dataGridWrapper.headers.isColumnHidden(0), 'first column is hidden');
-        assert.ok(dataGridWrapper.headers.isColumnHidden(1), 'second column is hidden');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'third column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(3), 'fourth column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(4), 'adaptive column is shown');
-        assert.equal($visibleColumns.length, 5, 'only 2 columns are visible');
-        assert.equal($visibleColumns.eq(2).text(), 'Age', 'First is \'age\' column');
-        assert.equal($visibleColumns.eq(3).text(), 'Country', 'Second is \'country\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns()[0].dataField, 'firstName', '\'firstName\' column is hidden');
-        assert.equal(adaptiveColumnsController.getHiddenColumns()[1].dataField, 'lastName', '\'lastName\' column is hidden');
-        assert.equal(adaptiveColumnsController.getHidingColumnsQueue().length, 2, 'There is no columns in hiding queue');
-
-        $('#container').width(900);
-        instance.updateDimensions();
-
-        this.clock.tick();
-
-        $visibleColumns = $(instance.$element().find('.dx-header-row td'));
-
-        // assert
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(0), 'first column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(1), 'second column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(2), 'third column is shown');
-        assert.ok(!dataGridWrapper.headers.isColumnHidden(3), 'fourth column is shown');
-        assert.ok(dataGridWrapper.headers.isColumnHidden(4), 'adaptive column is hidden');
-        assert.equal($visibleColumns.length, 5, '4 columns are visible');
-        assert.equal($visibleColumns.eq(0).text(), 'First Name', 'First is \'firstName\' column');
-        assert.equal($visibleColumns.eq(1).text(), 'Last Name', 'Second is \'lastName\' column');
-        assert.equal(adaptiveColumnsController.getHiddenColumns().length, 0, 'There is no hidden columns');
-        assert.equal(adaptiveColumnsController.getHidingColumnsQueue().length, 2, 'There is 2 columns in hiding queue');
-    });
-
-    QUnit.test('Form item of adaptive detail row is rendered with the underscore template', function(assert) {
-        // arrange
-        $('#container').width(200);
-
-        const data = [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy' }];
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columnHidingEnabled: true,
-            dataSource: data,
-            columns: ['firstName', { dataField: 'lastName', cellTemplate: $('#scriptTestTemplate1') }]
-        });
-        const instance = $dataGrid.dxDataGrid('instance');
-
-        // act
-        instance.expandAdaptiveDetailRow(data[0]);
-
-        // assert
-        assert.equal($dataGrid.find('.dx-adaptive-detail-row .dx-form').length, 1, 'adaptive detail form is opened');
-        assert.equal($dataGrid.find('.dx-form #template1').text(), 'Template1', 'the underscore template is rendered correctly');
-
-        instance.collapseAdaptiveDetailRow(data[0]);
-    });
-
-    QUnit.test('Get correct column and column index in the onCellHoverChanged event when event is occurred for form\'s item', function(assert) {
-        // arrange
-        $('#container').width(200);
-
-        const dataSource = [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy' }];
-        const eventArgs = [];
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columnHidingEnabled: true,
-            dataSource: dataSource,
-            columns: ['firstName', 'lastName'],
-            onCellHoverChanged: function(e) {
-                assert.equal(typeUtils.isRenderer(e.cellElement), !!config().useJQuery, 'cellElement is correct');
-                eventArgs.push({
-                    column: e.column,
-                    columnIndex: e.columnIndex
-                });
-            }
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-
-        // act
-        instance.expandAdaptiveDetailRow(dataSource[0]);
-        this.clock.tick();
-        dataGrid.find('.dx-field-item-content').first().trigger('mouseover');
-        dataGrid.find('.dx-field-item-content').first().trigger('mouseout');
-
-        // assert
-        assert.equal(eventArgs.length, 2, 'count of eventArgs');
-        assert.equal(eventArgs[0].column.dataField, 'lastName', 'dataField of column (mouseover)');
-        assert.equal(eventArgs[0].columnIndex, 1, 'index of column (mouseover)');
-        assert.equal(eventArgs[1].column.dataField, 'lastName', 'dataField of column (mouseover)');
-        assert.equal(eventArgs[1].columnIndex, 1, 'index of column (mouseover)');
-    });
-
-    QUnit.test('Get correct column and column index in the onCellClick event when event is occurred for form\'s item', function(assert) {
-        // arrange
-        $('#container').width(200);
-
-        const dataSource = [{ firstName: 'Blablablablablablablablablabla', lastName: 'Psy' }];
-        let column;
-        let columnIndex;
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columnHidingEnabled: true,
-            dataSource: dataSource,
-            columns: ['firstName', 'lastName'],
-            onCellClick: function(e) {
-                assert.equal(typeUtils.isRenderer(e.cellElement), !!config().useJQuery, 'cellElement is correct');
-                column = e.column;
-                columnIndex = e.columnIndex;
-            }
-        });
-        const instance = dataGrid.dxDataGrid('instance');
-
-        // act
-        instance.expandAdaptiveDetailRow(dataSource[0]);
-        this.clock.tick();
-        dataGrid.find('.dx-field-item-content').trigger('dxclick');
-
-        // assert
-        assert.equal(column.dataField, 'lastName', 'dataField of column');
-        assert.equal(columnIndex, 1, 'index of column');
-    });
-
     // T659247
     QUnit.test('Column widths for header cells should be correctly if columnAutoWidth is enabled and banded columns are used', function(assert) {
         // arrange
@@ -1888,135 +1663,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         });
     });
 
-    // T240338
-    QUnit.test('Loading columns state when all columns have width and one column is hidden', function(assert) {
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1', width: 100 }, { dataField: 'field2', width: 100 }, { dataField: 'field3', width: 100 }],
-            selection: {
-                mode: 'multiple'
-            },
-            columnChooser: { enabled: true },
-            dataSource: [],
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {
-                        columns: [{ dataField: 'field1', visibleIndex: 0, visible: true, width: 100 }, { dataField: 'field2', visibleIndex: 1, visible: true, width: 100 }, { dataField: 'field3', visibleIndex: 2, visible: false, width: 100 }]
-                    };
-                }
-            }
-
-        });
-
-        // assert
-        assert.equal(dataGrid.getController('columns').getVisibleColumns().length, 0, 'visible column count');
-
-        // act
-        this.clock.tick();
-
-        // assert
-        const visibleColumns = dataGrid.getController('columns').getVisibleColumns();
-        assert.equal(visibleColumns.length, 3, 'visible column count');
-        assert.equal(visibleColumns[0].command, 'select', 'select column');
-        assert.equal(visibleColumns[1].dataField, 'field1', 'field1 column');
-        assert.equal(visibleColumns[2].dataField, 'field2', 'field1 column');
-    });
-
-    // T862537
-    QUnit.test('column should be draggable if grid contains this column and column with allowHiding: false', function(assert) {
-        // act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columns: [{ dataField: 'field1', allowHiding: false }, { dataField: 'field2' }],
-            dataSource: []
-        });
-
-        // assert
-        assert.equal($(dataGrid.$element()).find('.dx-datagrid-drag-action').length, 0, 'no drag actions');
-        assert.equal($(dataGrid.$element()).find('.dx-datagrid-action').length, 2, 'two actions');
-
-        // act
-        dataGrid.showColumnChooser();
-
-        // assert
-        assert.equal($(dataGrid.$element()).find('.dx-datagrid-drag-action').length, 1, 'one drag action for hiding column');
-    });
-
-    QUnit.test('Correct runtime changing of a columnChooser mode (string)', function(assert) {
-        // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columns: [{ dataField: 'field1', allowSorting: false }, { dataField: 'field2' }],
-            dataSource: []
-        });
-
-        // act
-        dataGrid.showColumnChooser();
-
-        let $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        assert.ok($overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-drag'), 'has dragAndDrop mode class');
-        assert.ok(!$overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-select'), 'hasn\'t select mode class');
-
-        dataGrid.option('columnChooser.mode', 'select');
-
-        $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        // assert
-        assert.ok(!$overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-drag'), 'hasn\'t dragAndDrop mode class');
-        assert.ok($overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-select'), 'has select mode class');
-    });
-
-    QUnit.test('Correct runtime changing of a columnChooser mode (object)', function(assert) {
-        // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columns: [{ dataField: 'field1', allowSorting: false }, { dataField: 'field2' }],
-            dataSource: []
-        });
-
-        // act
-        dataGrid.showColumnChooser();
-
-        let $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        assert.ok($overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-drag'), 'has dragAndDrop mode class');
-        assert.ok(!$overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-select'), 'hasn\'t select mode class');
-
-        dataGrid.option({ columnChooser: { mode: 'select' } });
-
-        $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        // assert
-        assert.ok(!$overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-drag'), 'hasn\'t dragAndDrop mode class');
-        assert.ok($overlayWrapper.hasClass('dx-datagrid-column-chooser-mode-select'), 'has select mode class');
-    });
-
-    QUnit.test('ColumnChooser\'s treeView get correct default config (without checkboxes)', function(assert) {
-        // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columnChooser: { mode: 'select' },
-            columns: [{ dataField: 'field1', allowSorting: false }, { dataField: 'field2', visible: false }],
-            dataSource: []
-        });
-
-        // act
-        dataGrid.showColumnChooser();
-
-        let $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        assert.ok($overlayWrapper.find('.dx-checkbox').length, 'There are checkboxes in columnChooser');
-
-        dataGrid.option({ columnChooser: { mode: 'dragAndDrop' } });
-
-        $overlayWrapper = dataGrid.getView('columnChooserView')._popupContainer._wrapper();
-
-        // assert
-        assert.ok(!$overlayWrapper.find('.dx-checkbox').length, 'There aren\'t checkboxes in columnChooser');
-    });
-
     // T756338
     QUnit.test('keyOf should not be called too often after push with row updates', function(assert) {
         // arrange
@@ -2304,164 +1950,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         // assert
         assert.equal(toolbarItemOffset, $(dataGrid.$element()).find('.dx-datagrid-search-panel').offset().top, 'toolbar sarch panel is aligned');
         assert.equal(toolbarItemOffset, $(dataGrid.$element()).find('.dx-toolbar .dx-datebox').offset().top, 'toolbar custom item is aligned');
-    });
-
-    if(browser.msie && parseInt(browser.version) <= 11) {
-        QUnit.test('Update the scrollable for IE browsers when the adaptive column is hidden', function(assert) {
-            // arrange
-
-
-            const dataGrid = createDataGrid({
-                dataSource: [{
-                    'ID': 4,
-                    'OrderNumber': 35711,
-                    'OrderDate': '2014/01/12'
-                }],
-                columnAutoWidth: true,
-                columnHidingEnabled: true,
-                columns: ['ID', 'OrderNumber', 'OrderDate']
-            });
-
-            this.clock.tick();
-
-            // act
-            const scrollable = dataGrid.$element().find('.dx-scrollable').data('dxScrollable');
-            sinon.spy(scrollable, 'update');
-            dataGrid.updateDimensions();
-            this.clock.tick();
-
-            // assert
-            const $lastDataCell = dataGrid.$element().find('.dx-last-data-cell');
-            assert.equal($lastDataCell.text(), '2014/01/12', 'text of last data cell');
-            assert.equal(scrollable.update.callCount, 2);
-
-
-        });
-    }
-
-    QUnit.test('Error row should be shown when state loading failed (T894590)', function(assert) {
-        // arrange
-        const errorText = 'test error';
-        const contentReadyHandler = sinon.spy();
-        const dataErrorOccurred = sinon.spy();
-        const gridOptions = {
-            dataSource: [{ id: 1 }],
-            columns: ['id'],
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return $.Deferred().reject(errorText).promise();
-                }
-            },
-            onContentReady: contentReadyHandler,
-            onDataErrorOccurred: dataErrorOccurred
-        };
-        const dataGrid = createDataGrid(gridOptions);
-        this.clock.tick();
-
-        const $headerRow = $(dataGrid.element()).find('.dx-header-row');
-        const $errorRow = $(dataGrid.element()).find('.dx-error-row');
-        const renderedRowCount = dataGrid.getVisibleRows().length;
-
-        // assert
-        assert.ok(contentReadyHandler.called, 'onContentReady is called');
-        assert.equal(dataErrorOccurred.callCount, 1, 'onDataErrorOccurred is called');
-        assert.equal(dataErrorOccurred.getCall(0).args[0].error, errorText, 'error text is correct');
-        assert.equal(renderedRowCount, 0, 'there are no rendered data rows');
-        assert.ok($headerRow.length, 'header row is rendered');
-        assert.ok($errorRow.length, 'error row is rendered');
-        assert.equal($errorRow.find('.dx-error-message').text(), errorText, 'error text is correct');
-    });
-
-    QUnit.test('Error row should display the default error message when reject is called without a parameter in stateStoring.customLoad (T894590)', function(assert) {
-        // arrange
-        const gridOptions = {
-            dataSource: [],
-            columns: ['id'],
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return $.Deferred().reject().promise();
-                }
-            }
-        };
-        const dataGrid = createDataGrid(gridOptions);
-        this.clock.tick();
-
-        const $errorRow = $(dataGrid.element()).find('.dx-error-row');
-
-        // assert
-        assert.ok($errorRow.length, 'error row is rendered');
-        assert.equal($errorRow.find('.dx-error-message').text(), 'Unknown error', 'default error message');
-    });
-
-    QUnit.test('Error row should not be displayed when reject is called in stateStoring.customLoad and errorRowEnabled === false (T894590)', function(assert) {
-        // arrange
-        const dataErrorOccurred = sinon.spy();
-        const gridOptions = {
-            dataSource: [],
-            columns: ['id'],
-            errorRowEnabled: false,
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return $.Deferred().reject().promise();
-                }
-            },
-            onDataErrorOccurred: dataErrorOccurred
-        };
-        const dataGrid = createDataGrid(gridOptions);
-        this.clock.tick();
-
-        const $errorRow = $(dataGrid.element()).find('.dx-error-row');
-
-        // assert
-        assert.equal(dataErrorOccurred.callCount, 1, 'onDataErrorOccurred is called');
-        assert.equal(dataErrorOccurred.getCall(0).args[0].error, 'Unknown error', 'default error message');
-        assert.notOk($errorRow.length, 'error row is not rendered');
-    });
-
-    // T921829
-    QUnit.test('Row adding should work correctly if add button was clicked before table render', function(assert) {
-        // arrange
-        const dataGrid = createDataGrid({
-            dataSource: {
-                load: function() {
-                    const d = $.Deferred();
-                    setTimeout(() => {
-                        d.resolve([]);
-                    });
-
-                    return d;
-                }
-            },
-            columns: [{
-                dataField: 'field1',
-                fixed: true
-            }, 'field2', 'field3', 'field4', 'field5'],
-            showBorders: true,
-            editing: {
-                allowAdding: true
-            },
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {};
-                }
-            }
-        });
-
-        // act
-        $('.dx-datagrid-addrow-button').trigger('dxclick');
-        this.clock.tick();
-
-        // assert
-        const rows = dataGrid.getVisibleRows();
-        assert.equal(rows.length, 1, 'row was added');
     });
 });
 
@@ -3644,30 +3132,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.strictEqual($dataGrid.find('.dx-data-row').children().length, 2, 'data cells count');
     });
 
-    // T689294
-    QUnit.test('onContentReady when there is no dataSource and stateStoring is enabled', function(assert) {
-        // arrange
-        let contentReadyCallCount = 0;
-
-        // act
-        createDataGrid({
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {};
-                }
-            },
-            onContentReady: function() {
-                contentReadyCallCount++;
-            }
-        });
-        this.clock.tick();
-
-        // assert
-        assert.equal(contentReadyCallCount, 1);
-    });
-
     // T824018
     QUnit.test('The onOptionChanged event should be called once when changing column option', function(assert) {
         // arrange
@@ -4068,28 +3532,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(eventArray[2], 'onContentReady', 'onContentReady event fired after closing error row');
     });
 
-    // T726366
-    QUnit.test('Column hiding should works correctly if all columns have width', function(assert) {
-        // arrange, act
-        const dataGrid = createDataGrid({
-            width: 300,
-            columnWidth: 100,
-            loadingTimeout: undefined,
-            columnHidingEnabled: true,
-            dataSource: [{}],
-            columns: ['field1', 'field2', 'field3', 'field4']
-        });
-
-        // assert
-        const visibleWidths = dataGrid.getVisibleColumns().map(column => column.visibleWidth);
-
-        assert.deepEqual(visibleWidths.length, 5, 'column count');
-        assert.deepEqual(visibleWidths[0], 100, 'column 1 has full width');
-        assert.deepEqual(visibleWidths[1], 'auto', 'column 2 has auto width');
-        assert.deepEqual(visibleWidths[2], 'adaptiveHidden', 'column 3 is hidden');
-        assert.deepEqual(visibleWidths[3], 'adaptiveHidden', 'column 4 is hidden');
-    });
-
     // T172125
     QUnit.test('resize when all columns have width', function(assert) {
         // arrange, act
@@ -4197,157 +3639,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         // assert
         assert.ok($(dataGrid.$element()).width() < $('#qunit-fixture').width(), 'total width after change pageIndex');
-    });
-
-    QUnit.test('Clear state when initial options defined', function(assert) {
-        const dataGrid = createDataGrid({
-            columns: [{ dataField: 'field1', sortOrder: 'desc' }, { dataField: 'field2' }, { dataField: 'field3' }],
-            dataSource: [],
-            columnChooser: { enabled: true },
-            paging: {
-                pageSize: 10
-            },
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {
-                        columns: [{ dataField: 'field1', visibleIndex: 0, visible: true }, { dataField: 'field2', visibleIndex: 1, visible: true }, { dataField: 'field3', visibleIndex: 2, visible: false }],
-                        pageSize: 40
-                    };
-                }
-            }
-
-        });
-
-        // act
-        this.clock.tick();
-
-        // assert
-        let visibleColumns = dataGrid.getController('columns').getVisibleColumns();
-        assert.equal(visibleColumns.length, 2, 'visible column count');
-        assert.equal(visibleColumns[0].sortOrder, undefined, 'field1 sortOrder');
-        assert.equal(dataGrid.pageSize(), 40, 'page size');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick();
-
-        // assert
-        visibleColumns = dataGrid.getController('columns').getVisibleColumns();
-        assert.equal(visibleColumns.length, 3, 'visible column count');
-        assert.equal(visibleColumns[0].sortOrder, 'desc', 'field1 sortOrder');
-        assert.equal(visibleColumns[0].sortIndex, 0, 'field1 sortIndex');
-        assert.equal(dataGrid.pageSize(), 10, 'page size');
-    });
-
-    // T528181
-    QUnit.test('Change state when lookup column exists and remote data is used', function(assert) {
-        const createRemoteDataSource = function(data) {
-            return {
-                key: 'id',
-                load: function() {
-                    const d = $.Deferred();
-
-                    setTimeout(function() {
-                        d.resolve(data);
-                    }, 0);
-
-                    return d.promise();
-                }
-            };
-        };
-
-        const dataGrid = createDataGrid({
-            columns: [{
-                dataField: 'id',
-                lookup: {
-                    dataSource: createRemoteDataSource([{ id: 1, text: 'Test 1' }]),
-                    valueExpr: 'id',
-                    displayExpr: 'text'
-                }
-            }],
-            dataSource: createRemoteDataSource([{ id: 1 }])
-        });
-
-        // act
-        this.clock.tick(0);
-
-        // act
-        dataGrid.state({});
-        this.clock.tick(0);
-
-        // assert
-        const $firstCell = $($(dataGrid.$element()).find('.dx-data-row').eq(0).children().eq(0));
-        assert.equal($firstCell.text(), 'Test 1', 'Lookup text is correct');
-    });
-
-    QUnit.test('Clear state when initial options is defined in dataSource', function(assert) {
-        const dataGrid = createDataGrid({
-            columnChooser: { enabled: true },
-            columns: [{ dataField: 'field1' }, { dataField: 'field2' }, { dataField: 'field3' }],
-            dataSource: {
-                sort: [{ selector: 'field1', desc: true }],
-                pageSize: 10,
-                store: []
-            },
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return {
-                        columns: [{ dataField: 'field1', visibleIndex: 0, visible: true }, { dataField: 'field2', visibleIndex: 1, visible: true }, { dataField: 'field3', visibleIndex: 2, visible: false }],
-                        pageSize: 40
-                    };
-                }
-            }
-
-        });
-
-        // act
-        this.clock.tick();
-
-        // assert
-        let visibleColumns = dataGrid.getController('columns').getVisibleColumns();
-        assert.equal(visibleColumns.length, 2, 'visible column count');
-        assert.equal(visibleColumns[0].sortOrder, undefined, 'field1 sortOrder');
-        assert.equal(visibleColumns[0].sortIndex, undefined, 'field1 sortIndex');
-        assert.equal(dataGrid.pageSize(), 40, 'page size');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick();
-
-        // assert
-        visibleColumns = dataGrid.getController('columns').getVisibleColumns();
-        assert.equal(visibleColumns.length, 3, 'visible column count');
-        assert.equal(visibleColumns[0].sortOrder, 'desc', 'field1 sortOrder');
-        assert.equal(visibleColumns[0].sortIndex, 0, 'field1 sortIndex');
-        assert.equal(dataGrid.pageSize(), 10, 'page size');
-    });
-
-    QUnit.test('Reset pageIndex on clear state', function(assert) {
-        const dataGrid = createDataGrid({
-            columns: ['field1'],
-            dataSource: [{}, {}, {}],
-            paging: {
-                pageSize: 2
-            }
-        });
-
-        // act
-        this.clock.tick();
-        dataGrid.pageIndex(1);
-
-        // assert
-        assert.equal(dataGrid.pageIndex(), 1, 'pageIndex');
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataGrid.pageIndex(), 0, 'pageIndex');
     });
 
     // T296786
@@ -4878,43 +4169,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.strictEqual(cellPreparedArgs[2].data, newItems[1], 'cellPrepared 2 data is updated');
     });
 
-    // T699807
-    QUnit.test('Change dataSource array during state loading', function(assert) {
-        // arrange
-        const stateDeferred = $.Deferred();
-        const dataGrid = createDataGrid({
-            stateStoring: {
-                enabled: true,
-                type: 'custom',
-                customLoad: function() {
-                    return stateDeferred;
-                }
-            },
-            keyExpr: 'id',
-            loadingTimeout: undefined,
-            repaintChangesOnly: true,
-            dataSource: [
-                { id: 1, field1: 'test1', detail: 'detail1' },
-                { id: 2, field1: 'test2', detail: 'detail2' }
-            ],
-            columns: ['id', 'field1']
-        });
-
-        this.clock.tick();
-
-        // act
-        const newItems = [
-            { id: 1, field1: 'test1', detail: 'detail1' },
-            { id: 2, field1: 'test2', detail: 'updated' }
-        ];
-
-        dataGrid.option('dataSource', newItems);
-        stateDeferred.resolve({});
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows()[1].data.detail, 'updated', 'row 1 data is updated');
-    });
-
     QUnit.test('watch in cellPrepared should works after push', function(assert) {
         // arrange
         const dataGrid = createDataGrid({
@@ -5000,43 +4254,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(cellPreparedArgs[0].oldValue, 'test1', 'cell prepared oldValue');
     });
 
-    QUnit.test('push changes for adaptive row', function(assert) {
-        // arrange
-        const dataSource = new DataSource({
-            pushAggregationTimeout: 0,
-            store: {
-                type: 'array',
-                key: 'id',
-                data: [
-                    { id: 1, field1: 'test1' },
-                    { id: 2, field1: 'test2' },
-                    { id: 3, field1: 'test3' },
-                    { id: 4, field1: 'test4' }
-                ]
-            }
-        });
-        const dataGrid = createDataGrid({
-            width: 100,
-            columnWidth: 100,
-            columnHidingEnabled: true,
-            repaintChangesOnly: true,
-            loadingTimeout: undefined,
-            keyExpr: 'id',
-            dataSource: dataSource
-        });
-
-
-        dataGrid.expandAdaptiveDetailRow(2);
-
-        const $cell = $(dataGrid.getCellElement(2, 1));
-
-        // act
-        dataGrid.getDataSource().store().push([{ type: 'update', key: 2, data: { field1: 'test updated' } }]);
-
-        // assert
-        assert.strictEqual($cell.text(), 'test updated', 'field1 text is updated');
-    });
-
     QUnit.test('Refresh with changesOnly and summary', function(assert) {
         // arrange
         const dataSource = new DataSource({
@@ -5077,33 +4294,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.ok($updatedCellElements.eq(0).is($cellElements.eq(0)), 'first cell isn\'t changed');
         assert.notOk($updatedCellElements.eq(1).is($cellElements.eq(1)), 'second cell is changed');
         assert.strictEqual($updatedCellElements.eq(1).text(), 'Sum: 500', 'cell value is updated');
-    });
-
-    // T558189
-    QUnit.test('Band columns should be displayed correctly after state is reset', function(assert) {
-        // arrange
-        let columns;
-        const dataGrid = createDataGrid({
-            dataSource: [{ field1: 1, field2: 2, field3: 3, field4: 4 }],
-            paging: {
-                pageIndex: 0
-            },
-            customizeColumns: function() { },
-            columns: ['field1', 'field2', { caption: 'Band Column', columns: ['field3', 'field4'] }]
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.state(null);
-        this.clock.tick();
-
-        // assert
-        columns = dataGrid.getVisibleColumns(0).map(function(column) { return column.caption; });
-        assert.deepEqual(columns, ['Field 1', 'Field 2', 'Band Column'], 'columns of the first level');
-
-        columns = dataGrid.getVisibleColumns(1).map(function(column) { return column.caption; });
-        assert.deepEqual(columns, ['Field 3', 'Field 4'], 'columns of the second level');
     });
 
     // T709033
